refactor(effects): 优化效果堆栈相关的数据类型和资源管理
- 将添加和移除效果的数量类型由int改为ushort以节省内存
- 新增WorldEffectContainer.TryPopMoreList方法以优化效果列表弹出逻辑
- 在释放效果时重置effect.Data确保数据干净
- 优化效果移除流程，新增FreeEffect方法集中处理效果释放和容器移除
- 调整效果堆栈计数更新逻辑保证类型一致

+++ FLib/Sources/WorldCores/Effects/WorldAddEffectEvent.cs
@@ -9,6 +9,6 @@ namespace FLib.WorldCores.Effects
         public uint Id;
         public WorldEffect Effect;
         public WorldEntity AddedBy;
-        public int AddCount;
+        public ushort AddCount;
     }
 }
